user registration and verification

+++ notes-controller.js
@@ -1,8 +1,8 @@
 const chalk = require("chalk").default;
 const Note = require("./models/Note");
 
-async function addNote(title) {
-    await Note.create({ title });
+async function addNote(title, owner) {
+    await Note.create({ title, owner });
     console.log(chalk.bgGreen("Note was added"));
 }
 
@@ -11,14 +11,18 @@ async function getNotes() {
     return notes;
 }
 
-async function updateNote(noteData) {
-    await Note.updateOne({ _id: noteData.id }, { title: noteData.title })
-    console.log(chalk.bgYellow(`Note with id='${noteData.id}' has been updated`));
+async function removeNote(id, owner) {
+    await Note.deleteOne({ _id: id, owner });
+    console.log(chalk.bgRed("Note removed"));
 }
 
-async function removeNote(id) {
-    await Note.deleteOne({ _id: id });
-    console.log(chalk.bgRed("Note removed"));
+async function updateNote(noteData, owner) {
+    const result = await Note.updateOne({ _id: noteData.id, owner }, { title: noteData.title })
+
+    if (result.matchedCount === 0) {
+        throw new Error('No note to edit')
+    }
+    console.log(chalk.bgYellow(`Note with id='${noteData.id}' has been updated`));
 }
 
 module.exports = {
